requirements.txt
celery==4.3.0
ckanapi==4.3
flask-marshmallow==0.11.0 
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask==1.1.1
gunicorn==20.0.4
psycopg2-binary==2.8.4
python-dotenv==0.12.0
redis==3.4.1
sendgrid-python==0.1.1
